Add MVC support for resolving request body bounds on generic parameters
Before this commit, Spring MVC was unable to resolve upper or lower
bounds on Spring MVC request body parameter when generic controller
inheritance was involved. This use case is common with Kotlin
which uses declaration-site variance for List<T>.

This commit adds a related GenericTypeResolver#resolveGenericBounds
used in HttpEntityMethodProcessor and RequestResponseBodyMethodProcessor
to resolve properly generic parameters.

Closes gh-24033